server/requirements.txt
python-socketio
fastapi
uvicorn[standard]